Fix broken function
